Feat: Add camera meal scanner with EU nutritional data
- Add capture-then-analyze camera flow (take photo -> review -> analyze)
- Integrate Gemini Vision API with multi-model fallback for quota handling
- Configure EU/Italian nutritional data instead of US values
- Add mock meals database for API-free fallback
- Fix Pydantic model to accept float values for macro tracking

+++ models.py
@@ -54,8 +54,8 @@ class DailyQuest(BaseModel):
     completed: bool
 
 class Macro(BaseModel):
-    current: int
-    target: int
+    current: float
+    target: float
 
 class Macros(BaseModel):
     calories: Macro
